delete ArticleViewHolder, added some logic and ui for ArticleDetailsFragment

+++ app/src/main/java/com/example/stakanchik/data/models/ArticlesDto.kt
@@ -1,7 +1,7 @@
 package com.example.stakanchik.data.models
 
 data class ArticlesDto(
-    val article_id: Long,
+    val article_id: Int,
     val topic: String,
     val text: String,
     val publish_date: String,
