+++ app/src/main/java/com/example/stakanchik/data/models/ArticlesEntity.kt
@@ -7,7 +7,7 @@ import java.util.*
 @Entity
 data class ArticlesEntity(
     @PrimaryKey(autoGenerate = true)
-    val article_id: Long,
+    val article_id: Int,
     val topic: String,
     val text: String,
     val publish_date: String,
